feat(Template): Make manifest.json optional
Make manifest.json optional for WASM templates.

For Visual Studio 2022, the wizard should show an additional dialog
to ask whether the user wants a manifest.json file.

For `dotnet new unoapp` and `dotnet new unoapp-uwp-net6`, a parameter,
`--wasm-pwa-manifest` can be used.
