+++ src/SolutionTemplate/Uno.ProjectTemplates.Dotnet/content/unoapp-uwp-net6/.template.config/template.json
@@ -125,7 +125,13 @@
       "datatype": "bool",
       "description": "If specified, skips the automatic restore of the project on create.",
       "defaultValue": "false"
-    }
+    },
+    "wasm-pwa-manifest": {
+			"type": "parameter",
+			"datatype": "bool",
+			"description": "Enables the generation of a PWA manifest for WebAssembly",
+			"defaultValue": "false"
+		}
   },
   "primaryOutputs": [
     {
@@ -214,6 +220,13 @@
           "condition": "(!vscode)",
           "exclude": [
             ".vscode/**/*"
+          ]
+        },
+        {
+          "condition": "(!wasm-pwa-manifest)",
+          "exclude": [
+            "UnoQuickStart.Wasm/manifest.json",
+            "UnoQuickStart.Wasm/Assets/AppIcon-*"
           ]
         }
       ]
