+++ src/SolutionTemplate/UnoSolutionTemplate.WinUI.net6/UnoSolutionTemplate.WinUI.net6.csproj
@@ -252,6 +252,7 @@
     <Content Include="Skia.WPF\__TemplateIcon.png" />
     <Content Include="Wasm\Assets\SplashScreen.png" />
     <Content Include="Wasm\LinkerConfig.xml" />
+    <Content Include="Wasm\manifest.json" />
     <None Include="Wasm\Properties\launchSettings.json" />
     <Content Include="Wasm\WasmCSS\Fonts.css" />
     <Content Include="Wasm\WasmScripts\AppManifest.js" />
