+++ src/SolutionTemplate/UnoSolutionTemplate.Wizard.2022/Forms/UnoOptions.cs
@@ -12,10 +12,8 @@ using Microsoft.VisualStudio.Shell.Interop;
 
 namespace UnoSolutionTemplate.Wizard.Forms
 {
-	public partial class UnoOptions : Form
+	public partial class UnoOptions : UnoOptionsBaseForm
 	{
-		private IServiceProvider _serviceProvider;
-
 		public bool UseWebAssembly => checkWebAssembly.Checked;
 		public bool UseiOS => checkiOS.Checked;
 		public bool UseAndroid => checkAndroid.Checked;
@@ -35,79 +33,20 @@ namespace UnoSolutionTemplate.Wizard.Forms
 		}
 
 		public UnoOptions(IServiceProvider serviceProvider)
+			: base(serviceProvider)
 		{
-			ThreadHelper.ThrowIfNotOnUIThread();
-
 			InitializeComponent();
 
-			_serviceProvider = serviceProvider;
-
-			if (_serviceProvider.GetService(typeof(IUIHostLocale)) is IUIHostLocale2 hostLocale)
-			{
-				UIDLGLOGFONT[] array = (UIDLGLOGFONT[])(object)new UIDLGLOGFONT[1];
-				if (hostLocale.GetDialogFont(array) == 0)
-				{
-					Font = FontFromUIDLGLOGFONT(array[0]);
-				}
-			}
+			InitializeFont();
 
 			BaseTargetFramework.Items.Add(new TargetFrameworkOption { BaseValue = "net6.0", DisplayValue = ".NET 6.0" });
 			BaseTargetFramework.Items.Add(new TargetFrameworkOption { BaseValue = "net7.0", DisplayValue = ".NET 7.0" });
 			BaseTargetFramework.SelectedIndex = 0;
 		}
 
-		private static Font FontFromUIDLGLOGFONT(UIDLGLOGFONT logFont)
-		{
-			var fonts = new char[logFont.lfFaceName.Length];
-
-			var num = 0;
-			ushort[] lfFaceName = logFont.lfFaceName;
-			foreach (ushort num2 in lfFaceName)
-			{
-				fonts[num++] = (char)num2;
-			}
-
-			var familyName = new string(fonts);
-			var emSize = -logFont.lfHeight;
-			var fontStyle = FontStyle.Regular;
-
-			if (logFont.lfItalic > 0)
-			{
-				fontStyle |= FontStyle.Italic;
-			}
-			if (logFont.lfUnderline > 0)
-			{
-				fontStyle |= FontStyle.Underline;
-			}
-			if (logFont.lfStrikeOut > 0)
-			{
-				fontStyle |= FontStyle.Strikeout;
-			}
-			if (logFont.lfWeight > 400)
-			{
-				fontStyle |= FontStyle.Bold;
-			}
-
-			var unit = GraphicsUnit.Pixel;
-			var lfCharSet = logFont.lfCharSet;
-
-			return new Font(familyName, emSize, fontStyle, unit, lfCharSet);
-		}
-
 		private void UnoOptions_Load(object sender, EventArgs e)
 		{
-			using Graphics graphics = CreateGraphics();
-			var sizeF = graphics.MeasureString(labelDescription.Text, Font);
-
-			int widthRatio = (int)(sizeF.Width / (float)labelDescription.Width);
-			if (widthRatio != 0)
-			{
-				int heightCeil = (int)Math.Ceiling(sizeF.Height);
-				SuspendLayout();
-				labelDescription.Height = heightCeil + widthRatio * heightCeil;
-				ResumeLayout(performLayout: true);
-			}
-
+			ResizeLabelDescription(labelDescription);
 			CenterToParent();
 		}
