+++ src/SolutionTemplate/UnoSolutionTemplate.Wizard.2022/Forms/UnoWasmOptions.cs
@@ -1,0 +1,23 @@
+﻿using System;
+
+namespace UnoSolutionTemplate.Wizard.Forms
+{
+	public partial class UnoWasmOptions : UnoOptionsBaseForm
+	{
+		public bool UseManifestJson => checkManifestJson.Checked;
+
+		public UnoWasmOptions(IServiceProvider serviceProvider)
+			: base(serviceProvider)
+		{
+			InitializeComponent();
+
+			InitializeFont();
+		}
+
+		private void UnoWasmOptions_Load(object sender, EventArgs e)
+		{
+			ResizeLabelDescription(labelDescription);
+			CenterToParent();
+		}
+	}
+}
